pyproject.toml
[tool.poetry]
name = "pydantic_bigquery"
version = "0.4.0"
description = "BigQuery integration with Pydantic models"
authors = ["Hynek Blaha <hynek.blaha.dev@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7"
pydantic = "*"
structlog = "*"
google-cloud-bigquery = "^3.12.0"
backoff = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
mypy = "*"
pylint = "*"
pre-commit = "*"
Faker = "*"
black = "*"
isort = "*"
mock = "*"
types-mock = "*"
types-setuptools = "*"

[tool.black]
line-length = 120

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true
known_first_party = []
known_third_party = []

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
